Initial commit into this repository
Please note that the code was written back in 2000..2002 and it needs
some adjusting for it to work with modern Visual C++. I used
Visual C++ Express 2010 to convert the project from the old VC++ format.

I'm sharing this code for those that might be interested with COM
objects programming. However, nowadays I'm a Linux user and I no
longer have Visual Studio Professional suite, so be free to adjust
the code as needed.